Move keys dialog function into babonus
