+++ scripts/helpers/helpers.mjs
@@ -241,49 +241,3 @@ export function _getCollection(object) {
   }, []);
   return new foundry.utils.Collection(contents);
 }
-
-/**
- * Helper function to display the keys dialog and subsequently place the
- * selected values in the input fields that its button was placed near.
- * @param {PointerEvent} event      The initiating click event.
- */
-export async function _onDisplayKeysDialog(event) {
-  const formGroup = event.currentTarget.closest(".form-group");
-  const filterId = formGroup.dataset.id;
-
-  const lists = foundry.utils.duplicate(KeyGetter[filterId]);
-
-  // The text inputs.
-  const inputs = formGroup.querySelectorAll("input[type='text']");
-  const double = inputs.length === 2;
-
-  const [list, list2] = inputs
-  const values0 = inputs[0].value.split(";");
-  const values1 = inputs[1]?.value.split(";");
-  lists.forEach(t => {
-    t.checked0 = values0.includes(t.value);
-    t.checked1 = values1?.includes(t.value);
-  });
-  const selected = await BabonusKeysDialog.prompt({
-    label: game.i18n.localize("BABONUS.KeysDialogApplySelection"),
-    rejectClose: false,
-    options: {filterId, appId: this.appId, lists, double},
-    callback: function(html) {
-      const selector = "td:nth-child(2) input[type='checkbox']:checked";
-      const selector2 = "td:nth-child(3) input[type='checkbox']:checked";
-      const checked = [...html[0].querySelectorAll(selector)];
-      const checked2 = [...html[0].querySelectorAll(selector2)];
-      return {
-        first: checked.map(i => i.id).join(";") ?? "",
-        second: checked2.map(i => i.id).join(";") ?? ""
-      };
-    },
-  });
-
-  if (!selected) return;
-  if (Object.values(selected).every(a => foundry.utils.isEmpty(a))) return;
-
-  list.value = selected.first;
-  if (list2) list2.value = selected.second;
-  return;
-}
